+++ client/src/pages/Posting.tsx
@@ -56,6 +56,7 @@ import useNotifications from '../notifications/useNotifications';
 import { postingEditFormSchema, type PostingEditFormData } from '../schemas/posting';
 import { executeAndShowError, FormField } from '../utils/formUtils.tsx';
 import requestServer from '../utils/requestServer.ts';
+import { toLocalTime, toUtcTime } from '../utils/timeUtils.ts';
 import useAsync from '../utils/useAsync';
 import { useOrganization } from '../utils/useUsers.ts';
 
@@ -139,24 +140,6 @@ const getDateInputValue = (value: Date | string) => {
 };
 
 const getTimeInputValue = (timeValue: string | undefined) => (timeValue ?? '').slice(0, 5);
-
-const getLocalOffsetMinutes = () => -new Date().getTimezoneOffset();
-
-const toUtcTime = (localTime: string): string => {
-  const [hh, mm] = localTime.split(':').map(Number);
-  const totalMinutes = (hh * 60 + mm) - getLocalOffsetMinutes();
-  const utcHh = ((totalMinutes / 60 | 0) + 24) % 24;
-  const utcMm = ((totalMinutes % 60) + 60) % 60;
-  return `${String(utcHh).padStart(2, '0')}:${String(utcMm).padStart(2, '0')}`;
-};
-
-const toLocalTime = (utcTime: string): string => {
-  const [hh, mm] = utcTime.split(':').map(Number);
-  const totalMinutes = (hh * 60 + mm) + getLocalOffsetMinutes();
-  const localHh = ((totalMinutes / 60 | 0) + 24) % 24;
-  const localMm = ((totalMinutes % 60) + 60) % 60;
-  return `${String(localHh).padStart(2, '0')}:${String(localMm).padStart(2, '0')}`;
-};
 
 const getPostingStartDateTime = (posting: PostingWithSkills) => {
   const datePart = getDateInputValue(posting.start_date);
